Don't ask authorization for managed user creation
Creating a managed user is just an admin action to express the
_intention_ to impersonate someone in the future. It does not imply
making any action on behalf on anyone, that's what impersonation itself
is for. So we don't need to require any authorization here.

This greatly improves the workflow since previously to create a managed
user and impersonate her you would have to introduce the same
authorization information twice in a row.

It also fixes the weird error message happening when the user was
already being impersonated when creating the managed user. It can no
longer happen.